Use subunit for the unit tests
Subunit[1] is a test result streaming protocol. It allows us to
easily aggregate and parse test results.

We'll start by using it for the unit tests, afterwards adapting
the integration tests to use it (where it will be even more
useful, as it's quite difficult ATM to see which integration
tests have been run, and we're currently losing test results
in case of crashes).

[1] https://github.com/testing-cabal/subunit